Add files via upload
This is NOT a final cheat. This is a basic CSGO cheat SDK which is aimed to help newer people to coding internals for CSGO. 
This cheat SDK has examples of grabbing interfaces automatically, hooking to the engine and drawing text with surface in PaintTraverse. 
This is by no means professional and the initial commit was made in around 2 hours but nonetheless a great SDK to use and learn from.